Article Class, Home Page, Footer, Navigation
Added an Article class, it only pull text_content so far, need to implement the video and images later. Also need to add a new function to addArticle.

Created a layout for home page and make sure bootstrap working fine with twig etc.

Added a simple footer with copyright.

Modified the Navigation to fit container width, just so it looks better on smaller device.

NOTE: THE LAYOUT IS NOT WORKING ON MOBILE!!!